matrix: switch to double precision for all matrix calculations
Convert all internal arithmetic to 64-bit double precision. This all has
to be done at once to avoid massive breakage.

+++ globe.h
@@ -9,24 +9,24 @@
 struct globe {
 
 	// Cursor position in radians:
-	float lat;
-	float lon;
+	double lat;
+	double lon;
 
 	// Model matrix and components:
 	struct {
-		float model[16];
+		double model[16];
 		struct {
-			float lat[16];
-			float lon[16];
+			double lat[16];
+			double lon[16];
 		} rotate;
 	} matrix;
 
 	// Inverse matrices:
 	struct {
-		float model[16];
+		double model[16];
 		struct {
-			float lat[16];
-			float lon[16];
+			double lat[16];
+			double lon[16];
 		} rotate;
 	} invert;
 
@@ -40,5 +40,5 @@ extern const struct globe *globe_get (void);
 extern void globe_updated_reset (void);
 extern void globe_tile_to_sphere (const struct cache_node *n, struct cache_data *d);
 extern bool globe_intersect (const union vec *start, const union vec *dir, float *lat, float *lon);
-extern void globe_moveto (const float lat, const float lon);
+extern void globe_moveto (const double lat, const double lon);
 extern void globe_init (void);
